\name{bmem-package}
\alias{bmem-package}
\docType{package}
\title{Mediation analysis with missing data using bootstrap}
\description{
Four methods for mediation analysis with missing data: Listwise deletion, Pairwise deletion, Multiple imputation, and Two Stage Maximum Likelihood algorithm. For MI and TS-ML, auxiliary variables can be included. Bootstrap confidence intervals for mediation effects are obtained. The robust method is also implemented for TS-ML. Since version 1.4, bmem adds the capability to conduct power analysis for mediation models. Details about the methods used can be found in these articles. Zhang and Wang (2003) <doi:10.1007/s11336-012-9301-5>. Zhang (2014) <doi:10.3758/s13428-013-0424-0>.
}
\details{
\tabular{ll}{
Package: \tab bmem\cr
Type: \tab Package\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Zhiyong Zhang and Lijuan Wang

Maintainer: Zhiyong Zhang <zhiyongzhang@nd.edu>

}

\keyword{ bmem }

