\name{bmem.ci.bca1}
\alias{bmem.ci.bca1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
BCa for a single variable
}
\description{
BCa for a single variable
}
\usage{
bmem.ci.bca1(x, b, jack, cl = 0.95)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A vector from a bootstrap output.}
  \item{b}{Parameter estimate from the original sample}
  \item{jack}{A vector from a Jackknife analysis}
  \item{cl}{Confidence level. Default 0.95.}
}
