\name{bmem.em.cov}
\alias{bmem.em.cov}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Covariance matrix from EM
}
\description{
Covariance matrix from EM
}
\usage{
bmem.em.cov(xmis, moment = FALSE, max_it = 500)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{xmis}{An object from output of \code{\link{bmem.pattern}}.}
  \item{moment}{Whether estimating mean}
  \item{max_it}{Maximum number of iterations}
}
