\name{bmem.list.cov}
\alias{bmem.list.cov}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Covariance matrix for listwise deletion
}
\description{
Covariance matrix for listwise deletion
}
\usage{
bmem.list.cov(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A data set}
}
