\name{bmem.mi.cov}
\alias{bmem.mi.cov}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Covariance estimation for multiple imputation
}
\description{
Covariance estimation for multiple imputation
}
\usage{
bmem.mi.cov(x, m = 10, moment = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A data set}
  \item{m}{Number of imputations}
  \item{moment}{Estimate mean or not}
}
