\name{bmem.pair.cov}
\alias{bmem.pair.cov}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Covariance matrix estimation based on pairwise deletion}
\description{Covariance matrix estimation based on pairwise deletion}
\usage{
bmem.pair.cov(x, moment = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A data set}
  \item{moment}{Estimate mean or not}
}
