\name{bmem.pattern}
\alias{bmem.pattern}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Obtain missing data pattern information}
\description{Obtain missing data pattern information}
\usage{
bmem.pattern(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A data set}
}
