\name{bmem.raw2cov}
\alias{bmem.raw2cov}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert a raw moment matrix to covariance matrix
}
\description{
Convert a raw moment matrix to covariance matrix
}
\usage{
bmem.raw2cov(x)
}
\arguments{
  \item{x}{A moment matrix}
}

\value{A covariance matrix}
\author{
Zhiyong Zhang and Lijuan Wang
}
\references{
Zhang, Z. & Wang, L. (2011) Four methods for mediation analysis with missing data.

Zhang, Z. (2011) Robust mediation analysis with missing data and auxiliary variables.}
\seealso{
\code{\link{bmem}}, \code{\link{bmem.sobel}}, \code{\link{bmem.plot}}
}
