\name{bmem.sobel.ind}
\alias{bmem.sobel.ind}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Mediation analysis using sobel test for one indirect effect
}
\description{
Internal function
}
\usage{
bmem.sobel.ind(sem.object, ind)
}
\arguments{
  \item{sem.object}{A sem object}
  \item{ind}{Indirect effect}
}

\value{Internal output}
\author{
Zhiyong Zhang and Lijuan Wang
}
\references{
Zhang, Z. & Wang, L. (2011) Four methods for mediation analysis with missing data.

Zhang, Z. (2011) Robust mediation analysis with missing data and auxiliary variables.
}
\seealso{
\code{\link{bmem}}, \code{\link{bmem.sobel}}, \code{\link{bmem.plot}}
}
