\name{bmem.v}
\alias{bmem.v}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Select data according to a vector of indices
}
\description{
Select data according to a vector of indices
}
\usage{
bmem.v(x, v, moment = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A matrix}
  \item{v}{A vector of indices}
  \item{moment}{Covariane analysis or mean and covariance analysis}
}
