\name{popPar}
\alias{popPar}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get the population parameter values
}
\description{
Get the population parameter values including both direct and indirect effects in a model
}
\usage{
popPar(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A \code{\link{lavaan}} object
}
}

