\name{powermediation-package}
\alias{powermediation-package}
\docType{package}
\title{
Monte Carlo Based Statistical Power Analysis for Mediation Models with Excessive Skewness and Kurtosis
}
\description{
The package conducts power analysis for widely used mediation models. It implements a general method to conduct power analysis for almost any mediation models based on bootstrap methods. The method is based on Monte Carlo simulation but uses the bootstrap method to test mediation effect. The package also allows the specification of non-normal data in the Monte Carlo simulation. 
}

\author{
Zhiyong Zhang <johnnyzhz@gmail.com>
}
\references{
Zhiyong Zhang (2013). Monte Carlo Based Statistical Power Analysis for Mediation Models with Excessive Skewness and Kurtosis
}

\keyword{ package }
