\name{summary.power}
\alias{summary.power}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Organize the results into a table
}
\description{
This function is adpated from the \code{\link{lavaan}} summary function to put the results in a table. 
}
\usage{
\method{summary}{power}(object,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Output from the function either \code{\link{power.basic}} or \code{\link{power.boot}}.
}
\item{...}{Other options}
}
