\name{bmem-package}
\alias{bmem-package}
\docType{package}
\title{Mediation analysis with missing data using bootstrap}
\description{
Four methods for mediation analysis with missing data: Listwise deletion, Pairwise deletion, Multiple imputation, and Two-stage ML. For MI and TSML, auxiliary variables can be included. Bootstrap confidence intervals for mediation effects are obtained.
}
\details{
\tabular{ll}{
Package: \tab bmem\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-04\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Zhiyong Zhang and Lijuan Wang

Maintainer: Zhiyong Zhang <zhiyongzhang@nd.edu>

}

\keyword{ bmem }

