\name{bmem.cov}
\alias{bmem.cov}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate the covariance matrix based on a given ram model
}
\description{
Can be used to simulated data for an SEM model.
}
\usage{
bmem.cov(ram,obs.variables,moment=FALSE, debug=FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ram}{An ram model}
  \item{obs.variables}{Names of the observed variables}
  \item{moment}{Whether to use the mean structure}
  \item{debug}{debug mode}
}
