\name{bmem.list.jack}
\alias{bmem.list.jack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Jackknife for listwise deletion
}
\description{
Jackknife for listwise deletion
}
\usage{
bmem.list.jack(x, ram, indirect, moment = FALSE, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A data set}
  \item{ram}{RAM path for the mediaiton model}
  \item{indirect}{A vector of indirect effec}
  \item{moment}{Select mean structure or covariance analysis. moment=FALSE, covariance analysis. moment=TRUE, mean and covariance analysis.} 
  \item{\dots}{Other options for \code{\link{sem}} function can be used.}
}
